fix: refactor dimension validation into reusable concern
Extract validation logic for dimensions into a new `ValidateDimensionsConcern` module to enforce DRY principles.
Updated `Bench` and `RequestDistribution` models to include the concern and adjusted related test cases accordingly.
Localized error messages have also been updated for consistency.
